+++ src/generators/legacy-json/utils/buildSection.mjs
@@ -1,14 +1,7 @@
-import {
-  DEFAULT_EXPRESSION,
-  LEADING_HYPHEN,
-  NAME_EXPRESSION,
-  RETURN_EXPRESSION,
-  TYPE_EXPRESSION,
-} from '../constants.mjs';
 import { buildHierarchy } from './buildHierarchy.mjs';
-import parseSignature from './parseSignature.mjs';
 import { getRemarkRehype } from '../../../utils/remark.mjs';
 import { transformNodesToString } from '../../../utils/unist.mjs';
+import { parseList } from './parseList.mjs';
 
 const sectionTypePlurals = {
   module: 'modules',
@@ -72,77 +65,6 @@ function createSection(entry, head) {
 }
 
 /**
- *
- * @param {String} string
- * @returns {String}
- */
-function transformTypeReferences(string) {
-  // console.log(string)
-  return string.replaceAll(/`<([^>]+)>`/g, '{$1}').replaceAll('} | {', '|');
-}
-
-/**
- * Parses a list item to extract properties.
- * @param {import('mdast').ListItem} child - The list item node.
- * @returns {import('../types.d.ts').List} The parsed list.
- */
-function parseListItem(child) {
-  const current = {};
-
-  /**
-   * Extracts a pattern from text and assigns it to the current object.
-   * @param {string} text
-   * @param {RegExp} pattern
-   * @param {string} key
-   * @returns {string}
-   */
-  const extractPattern = (text, pattern, key) => {
-    const [, match] = text.match(pattern) || [];
-    if (match) {
-      current[key] = match.trim().replace(/\.$/, '');
-      return text.replace(pattern, '');
-    }
-    return text;
-  };
-
-  // Combine and clean text from child nodes, excluding nested lists
-  current.textRaw = transformTypeReferences(
-    transformNodesToString(
-      child.children.filter(node => node.type !== 'list'),
-      true
-    )
-      .replace(/\s+/g, ' ')
-      .replace(/<!--.*?-->/gs, '')
-  );
-
-  let text = current.textRaw;
-
-  // Determine if the current item is a return statement
-  if (RETURN_EXPRESSION.test(text)) {
-    current.name = 'return';
-    text = text.replace(RETURN_EXPRESSION, '');
-  } else {
-    text = extractPattern(text, NAME_EXPRESSION, 'name');
-  }
-
-  // Extract type and default values if present
-  text = extractPattern(text, TYPE_EXPRESSION, 'type');
-  text = extractPattern(text, DEFAULT_EXPRESSION, 'default');
-
-  // Assign the remaining text as the description after removing leading hyphens
-  current.desc = text.replace(LEADING_HYPHEN, '').trim() || undefined;
-
-  // Recursively parse nested options if a list is found within the list item
-  const optionsNode = child.children.find(child => child.type === 'list');
-
-  if (optionsNode) {
-    current.options = optionsNode.children.map(parseListItem);
-  }
-
-  return current;
-}
-
-/**
  * Parses stability metadata and adds it to the section.
  * @param {import('../types.d.ts').Section} section - The section to add stability to.
  * @param {Array} nodes - The AST nodes.
@@ -154,42 +76,6 @@ function parseStability(section, nodes, entry) {
     section.stability = json.index;
     section.stabilityText = json.description;
     nodes.splice(0, 1);
-  }
-}
-
-/**
- * Parses a list and updates the section accordingly.
- * @param {import('../types.d.ts').Section} section - The section to update.
- * @param {Array} nodes - The AST nodes.
- */
-function parseList(section, nodes) {
-  const list = nodes[0]?.type === 'list' ? nodes.shift() : null;
-
-  const values = list ? list.children.map(parseListItem) : [];
-
-  switch (section.type) {
-    case 'ctor':
-    case 'classMethod':
-    case 'method':
-      section.signatures = [parseSignature(section.textRaw, values)];
-      break;
-    case 'property':
-      if (values.length) {
-        const { type, ...rest } = values[0];
-
-        section.type = type;
-        Object.assign(section, rest);
-        section.textRaw = `\`${section.name}\` ${section.textRaw}`;
-      }
-      break;
-    case 'event':
-      section.params = values;
-      break;
-    default:
-      // If the list wasn't processed, add it back for further processing
-      if (list) {
-        nodes.unshift(list);
-      }
   }
 }
