+++ src/generators/legacy-json/utils/parseSignature.mjs
@@ -77,7 +77,7 @@ function parseDefaultValue(parameterName) {
 /**
  * @param {string} parameterName
  * @param {number} index
- * @param {Array<import('../types.d.ts').List>} markdownParameters
+ * @param {Array<import('../types.d.ts').ParameterList>} markdownParameters
  * @returns {import('../types.d.ts').Parameter}
  */
 function findParameter(parameterName, index, markdownParameters) {
@@ -108,7 +108,7 @@ function findParameter(parameterName, index, markdownParameters) {
 
 /**
  * @param {string[]} declaredParameters
- * @param {Array<import('../types.d.ts').List>} parameters
+ * @param {Array<import('../types.d.ts').ParameterList>} parameters
  */
 function parseParameters(declaredParameters, markdownParameters) {
   /**
@@ -165,7 +165,7 @@ function parseParameters(declaredParameters, markdownParameters) {
 
 /**
  * @param {string} textRaw Something like `new buffer.Blob([sources[, options]])`
- * @param {Array<import('../types.d.ts').List} markdownParameters The properties in the AST
+ * @param {Array<import('../types.d.ts').ParameterList} markdownParameters The properties in the AST
  * @returns {import('../types.d.ts').MethodSignature | undefined}
  */
 export default (textRaw, markdownParameters) => {
